Repository

+++ Training_BE_PS/src/main/java/io/interfaz/training/pojos/Orders.java
@@ -4,6 +4,7 @@
 package io.interfaz.training.pojos;
 
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Column;
@@ -11,6 +12,8 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
@@ -37,10 +40,11 @@ public class Orders {
 	@Column(name ="purchase_date")
 	private Date purchaseDate;
 	
-	private double subtotal;
+	private BigDecimal subtotal;
 	
-	private double iva;
+	private BigDecimal iva;
 	
-	private double total;
+	private BigDecimal total;
+	
 	
 }
