+++ Training_BE_PS/src/main/java/io/interfaz/training/pojos/Products.java
@@ -3,12 +3,13 @@
  */
 package io.interfaz.training.pojos;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.Table;
-
-import org.springframework.data.annotation.Id;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -18,23 +19,23 @@ import lombok.NoArgsConstructor;
  * @author tthornton
  *
  */
-@Table(name="products")
+@Table(name = "products")
 @Entity
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Products {
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private int id;
-	
+
 	private String name;
-	
+
 	private String description;
-	
-	private double price;
-	
+
+	private BigDecimal price;
+
 	private String status;
 
 }
